[WIP] Split code with config.rs
